library management system
Hence after the completion of the project we got familiar with the C programming and its features. A complete and useful library management can only be developed with lot of intensive effort and time. Due to lack of time and we are beginners in programming program that we expected can't be developed by us. Our  library management may not be most useful for library in our college but it will be the most useful for study  and programming practice using C.
As a whole, the project has been a good learning experience for us. We have gained knowledge about the various aspects of C programming. At the same time, we have developed a deep understanding about the handling in C.
We still want to emphasize that the program is not complete by itself. There is still a lot of room improvement. Graphics may be added to program to make it more attractive. The mouse cursor may be  initialized in order to make the program even more interactive.